requirements.txt
azure-functions
requests
azure-cosmos==4.2.0